[level 1] Title: 추억 점수, Time: 2.13 ms, Memory: 92.1 MB -BaekjoonHub

+++ 프로그래머스/1/176963. 추억 점수/추억 점수.java
@@ -10,11 +10,11 @@ class Solution {
         }
         
         for(int i = 0; i < photo.length ; i++){
-            Integer sum = 0;
+            int sum = 0;
             for(String p : photo[i]){
                 sum += data.get(p) != null ? data.get(p) : 0;
             }
-            result[i] = (int) sum;
+            result[i] = sum;
         }
